remove dart:io from public API
