+++ examples/flutter_gallery/lib/demo/all.dart
@@ -14,4 +14,3 @@ export 'pesto_demo.dart';
 export 'shrine_demo.dart';
 export 'transformations/transformations_demo.dart';
 export 'typography_demo.dart';
-export 'video_demo.dart';
