+++ packages/flutter/lib/src/foundation/_exit_web.dart
@@ -1,0 +1,9 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// The dart:html implementation of an exit call.
+///
+/// This is a no-op, as web applications do not have sufficient permissions to
+/// exit their own browser.
+void exit() { }
