+++ packages/flutter/lib/src/painting/debug.dart
@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:io';
-
 import 'package:flutter/foundation.dart';
 
 /// Whether to replace all shadows with solid color blocks.
@@ -16,7 +14,7 @@ bool debugDisableShadows = false;
 /// Signature for a method that returns an [HttpClient].
 ///
 /// Used by [debugNetworkImageHttpClientProvider].
-typedef HttpClientProvider = HttpClient Function();
+typedef HttpClientProvider = Object Function();
 
 /// Provider from which [NetworkImage] will get its [HttpClient] in debug builds.
 ///
