+++ packages/flutter/lib/src/services/asset_bundle.dart
@@ -4,11 +4,12 @@
 
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
 
+import '_network_asset_bundle_io.dart'
+  if (dart.library.html) '_network_asset_bundle_web.dart' as network_asset_bundle;
 import 'binary_messenger.dart';
 
 /// A collection of resources used by the application.
@@ -97,50 +98,14 @@ abstract class AssetBundle {
 
 /// An [AssetBundle] that loads resources over the network.
 ///
+/// This class is not supported in Flutter for web applications.
+///
 /// This asset bundle does not cache any resources, though the underlying
 /// network stack may implement some level of caching itself.
-class NetworkAssetBundle extends AssetBundle {
+abstract class NetworkAssetBundle extends AssetBundle {
   /// Creates an network asset bundle that resolves asset keys as URLs relative
   /// to the given base URL.
-  NetworkAssetBundle(Uri baseUrl)
-    : _baseUrl = baseUrl,
-      _httpClient = HttpClient();
-
-  final Uri _baseUrl;
-  final HttpClient _httpClient;
-
-  Uri _urlFromKey(String key) => _baseUrl.resolve(key);
-
-  @override
-  Future<ByteData> load(String key) async {
-    final HttpClientRequest request = await _httpClient.getUrl(_urlFromKey(key));
-    final HttpClientResponse response = await request.close();
-    if (response.statusCode != HttpStatus.ok)
-      throw FlutterError(
-        'Unable to load asset: $key\n'
-        'HTTP status code: ${response.statusCode}'
-      );
-    final Uint8List bytes = await consolidateHttpClientResponseBytes(response);
-    return bytes.buffer.asByteData();
-  }
-
-  /// Retrieve a string from the asset bundle, parse it with the given function,
-  /// and return the function's result.
-  ///
-  /// The result is not cached. The parser is run each time the resource is
-  /// fetched.
-  @override
-  Future<T> loadStructuredData<T>(String key, Future<T> parser(String value)) async {
-    assert(key != null);
-    assert(parser != null);
-    return parser(await loadString(key));
-  }
-
-  // TODO(ianh): Once the underlying network logic learns about caching, we
-  // should implement evict().
-
-  @override
-  String toString() => '${describeIdentity(this)}($_baseUrl)';
+  factory NetworkAssetBundle(Uri baseUrl) = network_asset_bundle.NetworkAssetBundle;
 }
 
 /// An [AssetBundle] that permanently caches string and structured resources
